% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_viewer-deps.R
\name{ortho_viewer_dependencies}
\alias{ortho_viewer_dependencies}
\title{htmlwidgets dependencies for the ortho_viewer widget}
\usage{
ortho_viewer_dependencies()
}
\description{
Returns NULL since all dependencies are now specified in the
ortho_viewer.yaml file (neuroimjs bundle, CSS, etc.).
}
\keyword{internal}
