% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_viewer.R
\name{ortho_viewerOutput}
\alias{ortho_viewerOutput}
\title{Shiny output for \code{ortho_viewer}}
\usage{
ortho_viewerOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Shiny output id.}

\item{width, height}{Output dimensions.}
}
\description{
Shiny output for \code{ortho_viewer}
}
