% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_events.R
\name{ortho_voxel}
\alias{ortho_voxel}
\title{Get voxel indices from an ortho_viewer event}
\usage{
ortho_voxel(event)
}
\arguments{
\item{event}{An ortho_event object or the raw input value.}
}
\value{
Integer vector of length 3 (i, j, k) as voxel indices,
or NULL if not available.
}
\description{
Get voxel indices from an ortho_viewer event
}
