% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_events.R
\name{print.ortho_event}
\alias{print.ortho_event}
\title{Print method for ortho_event objects}
\usage{
\method{print}{ortho_event}(x, ...)
}
\arguments{
\item{x}{An ortho_event object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for ortho_event objects
}
