% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_viewer.R
\name{renderOrtho_viewer}
\alias{renderOrtho_viewer}
\title{Shiny render function for \code{ortho_viewer}}
\usage{
renderOrtho_viewer(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns an \code{ortho_viewer} widget.}

\item{env, quoted}{Standard \code{htmlwidgets} arguments.}
}
\description{
Shiny render function for \code{ortho_viewer}
}
