% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ortho.R
\name{set_title}
\alias{set_title}
\title{Set viewer title}
\usage{
set_title(builder, title)
}
\arguments{
\item{builder}{An \code{ortho_builder} object.}

\item{title}{Character: the window/page title.}
}
\value{
The modified \code{ortho_builder} object (for chaining).
}
\description{
Set viewer title
}
